+++ src/views/CreditView.js
@@ -1,6 +1,6 @@
 import React from 'react';
 // import Balance from '../components/Balance';
-import CreditPage from '../components/Credit/CreditPage';
+import CreditPage from '../components/CreditPage';
 
 export default function BalanceView() {
     return (
