+++ src/views/DebitView.js
@@ -1,7 +1,7 @@
 import React from 'react';
 // import Balance from '../components/Balance';
 
-import DebitPage from '../components/Debit/DebitPage';
+import DebitPage from '../components/DebitPage';
 
 export default function DebitView() {
     return (
